+++ daisybell/__main__.py
@@ -7,7 +7,7 @@ from transformers import pipeline
 from tabulate import tabulate
 import pandas as pd
 
-from daisybell import scan
+from daisybell.daisybell import scan
 
 
 def create_file_output(
