+++ daisybell/scanners/ner_language_bias.py
@@ -1,0 +1,103 @@
+import json
+import os
+import tarfile
+from pathlib import Path
+from statistics import mean
+from typing import Any, Callable, Generator, Iterator, List, Sequence, Tuple
+from urllib.request import urlretrieve
+
+import pandas as pd
+import pysbd
+from tqdm.auto import tqdm
+from transformers import Pipeline, pipeline
+
+from daisybell.daisybell import scanner
+from daisybell.helpers.common import handle_common_params_to_masking_and_zeroshot
+from daisybell.helpers.dataset import emit_books, replace_entities
+
+@scanner(
+    name="ner-human-language-bias",
+    kind="bias",
+    description="Scanning for language bias in NER based models. WARNING! THIS SCANNER IS EXPERIMENTAL.",
+)
+class NerLanguageBias:
+    def can_scan(self, model: Pipeline) -> bool:
+        try:
+            return model.task == "token-classification"
+        except:
+            return False
+
+    def scan(self, model: Pipeline, params: dict) -> pd.DataFrame:
+        (
+            suffix,
+            max_names_per_language,
+            wikidata,
+        ) = handle_common_params_to_masking_and_zeroshot(params)
+
+        max_sentences_per_book = params.get("max_sentences_per_book", 999999999)
+        max_books = params.get("max_books", 999999999)
+
+        language_names = {}
+        for language in wikidata["language"].unique():
+            # get the names for the current language (up to the maximum number specified)
+            names = wikidata[wikidata["language"] == language]["name"].iloc[
+                :max_names_per_language
+            ]
+            # filter out names that are not one word
+            names = names[names.str.count(" ") == 0]
+            # Some languages just don't have enough examples, this skips them
+            if len(names) < 10:
+                continue
+            language_names[language] = names
+
+        splitter = pysbd.Segmenter(language="en", clean=True)
+        scores = {}
+        control_score = 0
+        for book_idx, (book_name, book_content) in enumerate(emit_books(params), 1):
+            print(f"Scanning using {book_name}...")
+            text = splitter.segment(book_content)
+            for sent_idx, sentence in enumerate(
+                tqdm(text, total=len(text), desc="Sentences", unit="sent")
+            ):
+                # control ner output
+                control = []
+                for entity in model(sentence):
+                    if "PER" in entity["entity"]:
+                        control.append(entity)
+                control_score += len(control)
+
+                for language in language_names.keys():
+                    transformed = replace_entities(
+                        sentence,
+                        (
+                            (
+                                entity["start"],
+                                entity["end"],
+                                language_names[language]
+                                .sample(random_state=18 + i)
+                                .item()
+                                + suffix,
+                            )
+                            for i, entity in enumerate(control)
+                        ),
+                    )
+                    test_result = []
+                    for entity in model(transformed):
+                        if "PER" in entity["entity"]:
+                            test_result.append(entity)
+
+                    scores[language] = scores.get(language, 0) + len(test_result)
+
+                if sent_idx == max_sentences_per_book:
+                    break
+            if book_idx == max_books:
+                break
+
+        for k, v in scores.items():
+            scores[k] = v / control_score
+
+        return (
+            pd.DataFrame({"Language": scores.keys(), "Recall": scores.values()})
+            .sort_values("Recall")
+            .reset_index(drop=True)
+        )
