+++ pyproject.toml
@@ -32,13 +32,15 @@ pytest = { version = "7.2.1", optional = true }
 pytest-cov = { version = "4.0.0", optional = true }
 pytype = { version = "2023.2.17", optional = true }
 flake8 = { version = "6.0.0", optional = true }
+pdbpp = {version = "^0.10.3", extras = ["test"]}
 
 [tool.poetry.extras]
 test = [
 "pytest",
 "pytest-cov",
 "pytype",
-"flake8"
+"flake8",
+"pdbpp"
 ]
 
 [tool.poetry.scripts]
