+++ tests/test_scanner.py
@@ -1,13 +1,14 @@
-from daisybell import scan
+from daisybell.daisybell import scan
+from daisybell.scanners import MaskingLanguageBias, NerLanguageBias, ZeroShotLanguageBias
 from transformers import pipeline
 
 
 def test_scanning_masking_human_bias():
+    # mlb = MaskingLanguageBias()
     res = scan(
         pipeline(model="roberta-base"),
         params={"max_names_per_language": 10},
     )
-
     name, kind, _, df = list(res)[0]
     assert name == "masking-human-language-bias"
     assert kind == "bias"
